closes #86 | improved seeder + fixed table

+++ database/migrations/2020_07_08_104328_create_apartments_table.php
@@ -21,7 +21,7 @@ class CreateApartmentsTable extends Migration
             $table->unsignedTinyInteger('rooms_number');
             $table->unsignedTinyInteger('beds_number');
             $table->unsignedTinyInteger('bathrooms_number');
-            $table->unsignedTinyInteger('square_meters');
+            $table->unsignedSmallInteger('square_meters');
             $table->string('country');
             $table->string('region');
             $table->string('province');
